Make peer address finding into a function
